rename: io_tools -> utils & add print_run_time for  generate_gallery
